Refactor and make more generic
This now is very similar to what thunk already does. This differences
being that it uses an emitter to accomplish asynchronous actions
allowing you to registerer and deregister pieces of middleware
on the fly. This may dangerously confuse data-flow as it will be
hard to determine if a listener is on or off so managing side-effects
may become harder.